requirements.txt
streamlit~=1.18.1
pandas~=1.4.2
matplotlib~=3.5.1
numpy~=1.22.3
statsmodels~=0.13.5
plotly~=5.13.0
seaborn~=0.11.2
pillow~=9.0.1
scikit-learn~=1.0.2
firebase-admin~=6.1.0
python-dotenv~=0.21.1